requirements.txt
# Image Generation
argparse
opencv-python
numpy
Pillow

# AI Model
cognitive-service-vision-model-customization-python-samples
